requirements.txt
dask==0.11.0
decorator==4.0.10
networkx==1.11
numpy==1.11.1
pandas==0.18.1
Pillow==3.3.1
pydicom==0.9.9
python-dateutil==2.5.3
pytz==2016.6.1
scikit-image==0.12.3
six==1.10.0
toolz==0.8.0
